requirements.txt
# Web Scraping
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0

# MCP SDK
mcp==1.1.2

# Environment Configuration
python-dotenv==1.0.1

# HTTP Requests (for potential API usage)
requests==2.32.3

# Date/Time Handling
python-dateutil==2.9.0.post0

# Logging
colorlog==6.9.0

# Development Tools
ruff==0.8.4
